# Don't clone the whole graph on each trait impl connection during DCA